test: restore: rename testing package name
